Made changes based on PR Comments
 - Made get-server-config a command/
 - Fixed Readme to reflect changes.
